+++ src/render_3d/shader_fn.rs
@@ -3,7 +3,7 @@ use shader::*;
 
 pub fn default(data: PixelData, uniform: &SceneInfo) -> PixelOutput {
     const SHADOW_COLOR: f32 = 0.8;
-    let mut shadow = data.normal.dot_product(-uniform.light_direction) as f32;
+    let mut shadow = data.normal.dot_product(-uniform.light_direction);
 
     shadow = (1. - SHADOW_COLOR) * shadow + SHADOW_COLOR;
